refactor: move validate_pem_file to a separate file
