+++ src/dfx/src/lib/identity/identity_utils.rs
@@ -1,13 +1,8 @@
 use crate::lib::error::DfxResult;
-use dfx_core::error::identity::IdentityError;
-use dfx_core::error::identity::IdentityError::{UnsupportedKeyVersion, ValidatePemContentFailed};
 
 use anyhow::Context;
 use candid::Principal;
 use fn_error_context::context;
-use ic_agent::identity::BasicIdentity;
-use ic_agent::identity::PemError;
-use ic_agent::identity::Secp256k1Identity;
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum CallSender {
@@ -28,21 +23,4 @@ pub async fn call_sender(wallet: &Option<String>) -> DfxResult<CallSender> {
         CallSender::SelectedId
     };
     Ok(sender)
-}
-
-pub fn validate_pem_file(pem_content: &[u8]) -> Result<(), IdentityError> {
-    let secp_res =
-        Secp256k1Identity::from_pem(pem_content).map_err(|e| ValidatePemContentFailed(Box::new(e)));
-    if let Err(e) = secp_res {
-        let basic_identity_res = BasicIdentity::from_pem(pem_content);
-        match basic_identity_res {
-            Err(PemError::KeyRejected(rj)) if rj.description_() == "VersionNotSupported" => {
-                return Err(UnsupportedKeyVersion());
-            }
-            Err(_) => return Err(e),
-            _ => {}
-        }
-    }
-
-    Ok(())
 }
